Rename encode and decode modules
