+++ crates/cosmos/cosmos-wasm-relayer/src/encoding/components.rs
@@ -3,8 +3,8 @@ use hermes_encoding_components::impls::delegate::DelegateEncoding;
 use hermes_encoding_components::impls::encoded::ProvideEncodedBytes;
 use hermes_encoding_components::impls::schema::ProvideStringSchema;
 pub use hermes_encoding_components::traits::convert::ConverterComponent;
-pub use hermes_encoding_components::traits::decoder::DecoderComponent;
-pub use hermes_encoding_components::traits::encoder::EncoderComponent;
+pub use hermes_encoding_components::traits::decode::DecoderComponent;
+pub use hermes_encoding_components::traits::encode::EncoderComponent;
 pub use hermes_encoding_components::traits::schema::SchemaGetterComponent;
 pub use hermes_encoding_components::traits::types::encoded::EncodedTypeComponent;
 pub use hermes_encoding_components::traits::types::schema::SchemaTypeComponent;
