+++ crates/cosmos/cosmos-wasm-relayer/src/types/client_state.rs
@@ -1,7 +1,7 @@
 use hermes_cosmos_chain_components::types::tendermint::TendermintClientState;
 use hermes_encoding_components::traits::convert::{CanConvert, Converter};
-use hermes_encoding_components::traits::decoder::CanDecode;
-use hermes_encoding_components::traits::encoder::CanEncode;
+use hermes_encoding_components::traits::decode::CanDecode;
+use hermes_encoding_components::traits::encode::CanEncode;
 use hermes_encoding_components::traits::types::encoded::HasEncodedType;
 use hermes_protobuf_encoding_components::types::ViaAny;
 use hermes_wasm_client_components::types::client_state::WasmClientState;
