+++ crates/encoding/encoding-components/src/impls/delegate.rs
@@ -3,8 +3,8 @@ use core::marker::PhantomData;
 use cgp_core::prelude::{DelegateComponent, HasErrorType};
 
 use crate::traits::convert::Converter;
-use crate::traits::decoder::Decoder;
-use crate::traits::encoder::Encoder;
+use crate::traits::decode::Decoder;
+use crate::traits::encode::Encoder;
 use crate::traits::schema::SchemaGetter;
 use crate::traits::types::encoded::HasEncodedType;
 use crate::traits::types::schema::HasSchemaType;
