+++ crates/encoding/encoding-components/src/impls/return_encoded.rs
@@ -1,7 +1,7 @@
 use cgp_core::error::HasErrorType;
 
-use crate::traits::decoder::Decoder;
-use crate::traits::encoder::Encoder;
+use crate::traits::decode::Decoder;
+use crate::traits::encode::Encoder;
 use crate::traits::types::encoded::HasEncodedType;
 
 pub struct ReturnEncoded;
