+++ crates/encoding/encoding-components/src/impls/use_strategy.rs
@@ -2,8 +2,8 @@ use core::marker::PhantomData;
 
 use cgp_core::Async;
 
-use crate::traits::decoder::{CanDecode, Decoder};
-use crate::traits::encoder::{CanEncode, Encoder};
+use crate::traits::decode::{CanDecode, Decoder};
+use crate::traits::encode::{CanEncode, Encoder};
 
 pub struct EncodeUsingStrategy<Strategy>(pub PhantomData<Strategy>);
