+++ crates/encoding/encoding-components/src/traits/encode.rs
