+++ crates/encoding/encoding-components/src/traits/mod.rs
@@ -1,10 +1,10 @@
 pub mod convert;
+pub mod decode;
 pub mod decode_mut;
-pub mod decoder;
+pub mod encode;
 pub mod encode_and_decode;
 pub mod encode_and_decode_mut;
 pub mod encode_mut;
-pub mod encoder;
 pub mod has_encoding;
 pub mod schema;
 pub mod types;
