+++ crates/encoding/protobuf-encoding-components/src/impls/protobuf.rs
@@ -1,6 +1,6 @@
 use cgp_core::error::{CanRaiseError, HasErrorType};
-use hermes_encoding_components::traits::decoder::Decoder;
-use hermes_encoding_components::traits::encoder::Encoder;
+use hermes_encoding_components::traits::decode::Decoder;
+use hermes_encoding_components::traits::encode::Encoder;
 use hermes_encoding_components::traits::types::encoded::HasEncodedType;
 use prost::{DecodeError, Message};
