+++ crates/encoding/protobuf-encoding-components/src/impls/via_any.rs
@@ -1,8 +1,8 @@
 use core::marker::PhantomData;
 
 use cgp_core::prelude::*;
-use hermes_encoding_components::traits::decoder::Decoder;
-use hermes_encoding_components::traits::encoder::Encoder;
+use hermes_encoding_components::traits::decode::Decoder;
+use hermes_encoding_components::traits::encode::Encoder;
 use hermes_encoding_components::traits::types::encoded::HasEncodedType;
 
 use crate::impls::any::{DecodeAsAnyProtobuf, EncodeAsAnyProtobuf};
